requirements.txt
discord.py==2.4.0
setuptools==73.0.1